requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.14.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
astunparse==1.6.3
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.1.0
av==16.0.1
beautifulsoup4==4.13.4
bleach==6.3.0
blinker==1.9.0
boto3==1.40.74
botocore==1.40.74
Brotli==1.1.0
cachetools==6.2.0
cairocffi==0.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.3
dataclasses-json==0.6.7
datasets==4.4.1
decorator==5.2.1
deepface==0.0.95
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
elevenlabs==2.16.0
et_xmlfile==2.0.0
evaluate==0.4.6
face-alignment==1.4.1
fastapi==0.118.0
fer==22.3.0
ffmpeg==1.4
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
fire==0.7.1
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==5.49.0
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kaggle==1.7.4.5
keras==3.11.3
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.0.10
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.45.1
lxml==6.0.0
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
openai==1.100.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pcb-tools==0.1.6
pillow==12.0.0
platformdirs==4.5.0
plotly==6.2.0
polars==1.34.0
polars-runtime-32==1.34.0
pooch==1.8.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
pygerber==2.4.3
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==1.0.0
PyMsgBox==2.0.1
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
retina-face==0.0.17
retrying==1.4.2
rich==14.1.0
rsa==4.9.1
rubicon-objc==0.5.2
ruff==0.13.3
s3transfer==0.14.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==4.0.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==1.0.0
spotipy==2.25.1
SQLAlchemy==2.0.43
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.48.0
supervision==0.26.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
text-unidecode==1.3
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2025.9.30
tiktoken==0.11.0
timm==1.0.20
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.0
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.204
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
